Support for nested tables.
Previously we parsed in keys for a table as 'key1.key2' rather than
having key2 be a nested table in key1. This change correctly creates the nested tables see Test.IO.Huck.Parser.prop_lenses.
